requirements.txt
python-telegram-bot==11.00